Expanded DisplayQuestion + changed boxes design

+++ ClientWPF/ClientWPF/CreateRoom.xaml.cs
@@ -63,8 +63,8 @@ namespace ClientWPF
             };
 
             string json = JsonConvert.SerializeObject(createRoom, Formatting.Indented);
-            CreateRoomResponse createRoomResponse = Communicator.ManageSendAndGetData<CreateRoomResponse>(
-                json, clientStream, (int)Codes.CREATE_ROOM_CODE);
+            CreateRoomResponse createRoomResponse = 
+                Communicator.ManageSendAndGetData<CreateRoomResponse>(json, clientStream, Codes.CREATE_ROOM_CODE);
 
             if (createRoomResponse.Status == ERROR_CODE)
             {
@@ -73,7 +73,7 @@ namespace ClientWPF
                 mainWindow.Show();
             }
             else
-            { 
+            {
                 int roomId = createRoomResponse.RoomId;
 
                 RoomData roomData = new RoomData
